requirements.txt
accelerate==1.10.1
bcrypt==4.3.0
boto3==1.40.27
botocore==1.40.27
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
einops==0.8.1
filelock==3.13.1
fsspec==2024.6.1
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
megfile==4.2.4
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
paramiko==3.5.1
pillow==11.3.0
psutil==7.0.0
pycparser==2.23
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.7.33
requests==2.32.5
s3transfer==0.14.0
safetensors==0.6.2
six==1.17.0
sympy==1.13.1
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.55.4
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.5.0
zipp==3.23.0
datasets==4.0.0
sentencepiece==0.2.0
ptyprocess==0.7.0
protobuf==6.31.1
opencv-python==4.11.0.86
lpips==0.1.4
scikit-image==0.25.2